\name{Genotype_data}
\alias{Genotype_data}
\title{Example genotype data for SMUT}
\description{
Example genotype data for SMUT. It is a matrix with 100 rows and 200 columns. Each row is an individual; each column is a SNP.
}
\format{
  It is a matrix with 100 rows and 200 columns. Each row is an individual; each column is a SNP.
}
